Resurrected the queryObject method of Get
Brought back the method to query mongo based on a dynamic query object.
I see the combo of name, description and environment as a key.... it
makes other stuff a lot easier. Maybe think about refactoring if the
object ever becomes more complicated, but works fine for now and is
relatively simple.